requirements.txt
absl-py==1.4.0
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
array-record==0.4.0
arrow==1.3.0
astor==0.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
autograd==1.6.2
autograd-gamma==0.5.0
auton-survival @ git+https://github.com/autonlab/auton-survival.git@5dde465f7223601717abddc1d075e837707c403b
Babel==2.14.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
bleach==6.1.0
Brotli==1.1.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
colorlog==6.8.2
comm==0.2.2
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.10
dataclasses==0.6
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
dm-tree==0.1.8
ecos==2.0.13
et-xmlfile==1.1.0
etils==1.3.0
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
feather-format==0.4.1
flatbuffers==24.3.25
fonttools==4.50.0
formulaic==0.2.4
fqdn==1.5.1
future==1.0.0
gast==0.4.0
gin-config==0.5.0
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.63.0
graphlib_backport==1.1.0
greenlet==3.0.3
grpcio==1.62.1
h11==0.14.0
h5py==3.10.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.6
imageio==2.34.1
immutabledict==4.2.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
inflate64==1.0.0
interface-meta==1.3.0
ipykernel==6.29.4
ipython==8.12.3
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.24
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.4
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.4
kaggle==1.6.14
keras==2.11.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==18.1.1
lifelines==0.26.4
llvmlite==0.41.1
lxml==5.2.2
Mako==1.3.2
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.6
mistune==3.0.2
multivae==0.1.1
multivolumefile==0.2.3
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.3
nest-asyncio==1.6.0
nltk==3.8.1
notebook_shim==0.2.4
numba==0.58.1
numexpr==2.8.6
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
optuna==3.6.1
osqp==0.6.5
overrides==7.7.0
packaging==24.0
pandas==1.5.3
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.0
portalocker==2.8.2
prometheus_client==0.20.0
promise==2.3
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==3.19.6
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
py7zr==0.21.0
pyarrow==15.0.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybcj==1.0.2
pycocotools==2.0.7
pycox==0.2.3
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.17.2
pyparsing==3.1.2
pyppmd==1.1.0
pythae==0.1.2
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-pptx==0.6.23
python-slugify==8.0.4
pytz==2024.1
PyYAML==5.4.1
pyzmq==25.1.2
pyzstd==0.15.10
qdldl==0.1.7.post0
referencing==0.34.0
regex==2024.5.15
requests==2.31.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
rpy2==3.5.1
rsa==4.9
sacrebleu==2.2.0
scikit-learn==1.2.2
scikit-multilearn==0.2.0
scikit-survival==0.21.0
scipy==1.10.1
seaborn==0.13.2
Send2Trash==1.8.2
sentencepiece==0.2.0
seqeval==1.2.2
shap==0.41.0
six==1.16.0
sklearn-pandas==2.2.0
slicer==0.0.7
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
stack-data==0.6.3
SurvivalEVAL==0.2.6
SurvSet==0.2.6
tabulate==0.9.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.1
tensorflow-addons==0.21.0
tensorflow-datasets==4.9.0
tensorflow-estimator==2.11.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-metadata==1.13.0
tensorflow-model-optimization==0.8.0
tensorflow-probability==0.19.0
tensorflow-text==2.10.0
termcolor==2.4.0
terminado==0.18.1
text-unidecode==1.3
texttable==1.7.0
tf-keras==2.15.0
tf-models-official==2.10.1
tf-slim==1.1.0
threadpoolctl==3.4.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
torch==1.13.1
torch-scatter==2.1.2
torch-sparse==0.6.18
torchtuples==0.2.2
torchvision==0.14.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
typeguard==2.13.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.10.0
tzlocal==5.2
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.2
wrapt==1.16.0
xgboost==2.0.3
xgbse==0.2.3
XlsxWriter==3.2.0
zipp==3.18.1
